Rename: Commons 폴더명 → Common으로 변경 (#59)
- git에서 대소문자 변경 인식을 못해서 폴더명 자체를 바꾼 후 되돌림
