+++ client/src/components/PlanDetail/PlanElement/Checklist/ListItem.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react'
-import Icon from '../../../Commons/Icon'
+import Icon from '../../../Common/Icon'
 import styles from './ListItem.module.scss'
 
 type LitItem = {
