+++ client/src/pages/Planner.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react'
-import Icon from '../components/Commons/Icon'
+import Icon from '../components/Common/Icon'
 import styles from './Planner.module.scss'
 import getCurrentUserPlanner from '../apis/planner'
 import PlanElement from '../components/Planner/PlanElement'
